# [stm32][sdmmc]Optimize execution efficiency and find the right clock division