@wdfk-prog 2024-08-05
- 另外,CUBEMX初始化的SDMMC时钟跑在PLL2R上,应该是`200mhz`;
- 另一个可选择的是PLL2Q,是`480MHZ`
- 但是使用`HAL_RCCEx_GetPeriphCLKFreq(RCC_PERIPHCLK_SDMMC)`获取的只有`480mhz`
- 强制将获取时钟频率改为`200mhz`,测试读写速率并没有提高或降低.
- 不知道是什么原因,art-pi的原理图上没有注明CLK的电阻是哪一颗,没办法用逻辑分析仪测量频率是否设置正常.