@wdfk-prog 2024-08-05
- 还有SD的性能提不上了,不知道时钟是否达标了,没法测试;
- 还是因为其他原因导致